feat(repvovec-ci): add docs-gate policy helper and coverage tests
- Implemented detailed evaluation of documentation-gate policy in the repovec-ci crate
- Added extensive inline documentation with examples for DocsGatePlan and DocsGateReason
- Introduced unit tests covering docs-gate classification scenarios
- Simplified main.rs by refactoring plan computation with compute_plan helper
- Updated developers-guide.md with structured sections on local quality gates and CI policy helper
- Minor workflow update to fix bun version reference

This enables better CI gating based on documentation changes, enhancing repo quality and tooling.
