+++ .github/workflows/ci.yml
@@ -145,7 +145,7 @@ jobs:
       - name: Markdown lint
         if: steps.docs_gate_plan.outputs.docs_gate_required == 'true'
         run: make markdownlint
-      - uses: oven-sh/setup-bun@v2
+      - uses: oven-sh/setup-bun@0c5077e514
         if: steps.docs_gate_plan.outputs.nixie_required == 'true'
         with:
           bun-version: "1.3.12"
